+++ yazi-actor/src/input/mod.rs
@@ -1,1 +1,1 @@
-yazi_macro::mod_flat!(close complete escape show);
+yazi_macro::mod_flat!(close complete escape recall remember show);